refacto: Rename errorPage to errorScreen
