--- client/src/errorPage/components/index.ts
@@ -1,1 +1,0 @@
-export { ErrorPage } from "./ErrorPage";
