--- client/src/errorPage/index.ts
@@ -1,2 +1,0 @@
-export * as i18n from "./i18n";
-export { ErrorPage } from "./components";
