+++ client/src/errorScreen/components/index.ts
@@ -1,0 +1,1 @@
+export { ErrorScreen } from "./ErrorScreen";
