+++ client/src/errorScreen/i18n/index.ts
