+++ client/src/errorScreen/index.ts
@@ -1,0 +1,2 @@
+export * as i18n from "./i18n";
+export { ErrorScreen } from "./components";
